+++ Dziedziczenie/trojkat.h
@@ -10,8 +10,8 @@ public:
 	Trojkat(int a);
 	Trojkat();
 	~Trojkat();
-	int ObliczPole();
-	int ObliczObwod();
+	double ObliczPole();
+	double ObliczObwod();
 	void WypiszDane();
 
 protected:
